+++ src/components/forms/change-password/change-password.jsx
@@ -1,31 +1,27 @@
-import React from 'react'
-import {Form, Input, Button, message} from 'antd'
-import 'antd/dist/antd.css'
-import {useState} from 'react'
-import {SmartRequest} from '../../utils/utils'
-import {useHistory} from 'react-router-dom'
+import {Button, Form, Input, message} from 'antd'
+import React, {useState} from 'react'
+import {SmartRequest} from '../../../utils/utils'
 
 
-const SignUp = () => {
+const ChangePassword = () => {
     const [form] = Form.useForm()
-    const {getFieldError, validateFields} = form
+    const {getFieldError, validateFields, resetFields} = form
     const [isButtonDisabled, setIsButtonDisabled] = useState(true)
     const [formError, setFormError] = useState('')
     const [fieldsErrors, setFieldsErrors] = useState({})
-    const history = useHistory()
+
 
     const onFinish = (values) => {
         setFormError('')
-        SmartRequest.post(
-            'signup/',
-            {user: values},
-            {},
-            false,
-            false
+        SmartRequest.patch(
+            'change-password/',
+            values,
+            {}
         )
-            .then(() => {
-                message.success('Successfully created account')
-                history.push('/log-in')
+            .then(resp => {
+                console.log('success in change password:', resp)
+                message.success('Successfully changed password')
+                resetFields()
             })
             .catch(error => {
                 if (error.response && error.response.status === 400) {
@@ -33,10 +29,10 @@ const SignUp = () => {
                     if (typeof error.response.data['detail'] !== 'object') {
                         setFormError(error.response.data['detail'])
                     } else {
-                        setFieldsErrors(error.response.data['detail']['user'])
+                        setFieldsErrors(error.response.data['detail'])
                     }
                 } else {
-                    console.error('catch on sign up: ', error)
+                    console.error('catch on password change: ', error)
                 }
             })
     }
@@ -91,49 +87,33 @@ const SignUp = () => {
                 <span className="ant-form-item-explain ant-form-item-explain-error">{formError}</span>
             </Form.Item>
             <Form.Item
-                label="Username"
-                name="username"
-                validateStatus={fieldsErrors['username'] && fieldsErrors['username'].length ? 'error' : ''}
-                help={fieldsErrors['username'] && fieldsErrors['username'].length ? fieldsErrors['username'][0] : ''}
+                label="Old password"
+                name="password_old"
+                validateStatus={fieldsErrors['password_old'] && fieldsErrors['password_old'].length ? 'error' : ''}
+                help={fieldsErrors['password_old'] && fieldsErrors['password_old'].length ? fieldsErrors['password_old'][0] : ''}
                 rules={[
                     {
                         required: true,
-                        message: 'Please input your username!',
+                        message: 'Please input your old password!',
                     },
                 ]}
             >
-                <Input autoComplete="username"/>
-            </Form.Item>
-
-            <Form.Item
-                label="Email"
-                name="email"
-                validateStatus={fieldsErrors['email'] && fieldsErrors['email'].length ? 'error' : ''}
-                help={fieldsErrors['email'] && fieldsErrors['email'].length ? fieldsErrors['email'][0] : ''}
-                rules={[
-                    {
-                        required: true,
-                        message: 'Please input your email!',
-                    },
-                ]}
-            >
-                <Input type="email" autoComplete="email"/>
+                <Input.Password autoComplete="off"/>
             </Form.Item>
             <Form.Item
-                label="Password"
-                name="password"
-                validateStatus={fieldsErrors['password'] && fieldsErrors['password'].length ? 'error' : ''}
-                help={fieldsErrors['password'] && fieldsErrors['password'].length ? fieldsErrors['password'][0] : ''}
+                label="New password"
+                name="password_new"
+                validateStatus={fieldsErrors['password_new'] && fieldsErrors['password_new'].length ? 'error' : ''}
+                help={fieldsErrors['password_new'] && fieldsErrors['password_new'].length ? fieldsErrors['password_new'][0] : ''}
                 rules={[
                     {
                         required: true,
-                        message: 'Please input your password!',
+                        message: 'Please input your new password!',
                     },
                 ]}
             >
-                <Input.Password autoComplete="current-password"/>
+                <Input.Password autoComplete="off"/>
             </Form.Item>
-
             <Form.Item
                 wrapperCol={{
                     offset: 8,
@@ -141,12 +121,12 @@ const SignUp = () => {
                 }}
             >
                 <Button disabled={isButtonDisabled} type="primary" htmlType="submit">
-                    Sign up
+                    Change password
                 </Button>
             </Form.Item>
+
         </Form>
     )
 }
 
-
-export default SignUp
+export default ChangePassword
