+++ src/components/forms/log-in/log-in.jsx
@@ -3,8 +3,8 @@ import {Form, Input, Button, message} from 'antd'
 import 'antd/dist/antd.css'
 import {useState} from 'react'
 import {useDispatch} from 'react-redux'
-import {setCurrentUserAsync} from '../../redux/user/user.actions'
-import {SmartRequest} from '../../utils/utils'
+import {setCurrentUserAsync} from '../../../redux/user/user.actions'
+import {SmartRequest} from '../../../utils/utils'
 import {useHistory} from 'react-router-dom'
